refactor(api): split profileUpdate from user service
