+++ packages/api/src/modules/user/services/profile/user.profile.service.ts
@@ -5,13 +5,19 @@ import {
     AgentTypeEnum,
     ResetPasswordErrorCodes,
     TerritorialScopeEnum,
+    UpdatableUser,
+    UserDto,
 } from "dto";
 import { isInObjectValues } from "../../../../shared/Validators";
-import { BadRequestError } from "../../../../shared/errors/httpErrors";
+import { BadRequestError, UserNotFoundError } from "../../../../shared/errors/httpErrors";
 import { joinEnum } from "../../../../shared/helpers/ArrayHelper";
 import userCheckService from "../check/user.check.service";
 import { UserService } from "../../user.service";
 import { sanitizeToPlainText } from "../../../../shared/helpers/StringHelper";
+import userRepository from "../../repositories/user.repository";
+import { removeSecrets } from "../../../../shared/helpers/RepositoryHelper";
+import notifyService from "../../../notify/notify.service";
+import { NotificationType } from "../../../notify/@types/NotificationType";
 
 export class UserProfileService {
     validateUserProfileData(userInfo, withPassword = true): { valid: false; error: Error } | { valid: true } {
@@ -88,6 +94,22 @@ export class UserProfileService {
             if (field in unsafeUserInfo) sanitizedUserInfo[field] = sanitizeToPlainText(unsafeUserInfo[field]);
         });
         return sanitizedUserInfo;
+    }
+
+    async profileUpdate(user: UserDto, data: Partial<UpdatableUser>): Promise<UserDto> {
+        if (!user) throw new UserNotFoundError();
+
+        const toBeUpdatedUser = { ...user, ...data };
+
+        const userInfoValidation = userProfileService.validateUserProfileData(toBeUpdatedUser, false);
+        if (!userInfoValidation.valid) throw userInfoValidation.error;
+
+        const safeUserInfo = userProfileService.sanitizeUserProfileData(data);
+        const updatedUser = await userRepository.update({ ...user, ...safeUserInfo });
+
+        const safeUpdatedUser = removeSecrets(updatedUser);
+        notifyService.notify(NotificationType.USER_UPDATED, safeUpdatedUser);
+        return safeUpdatedUser;
     }
 }
